Refactor selection lookup logic
To make the code easier to maintain I'm trying to create some pipelines
that make it easy to understand how selections are found.
This will ultimately make the selection lookup process easier to
understand.

This has already paid off in the form of finding a serious oversight in
the assumptions that form the basis for how selections were detected;
when selections happen to be side-by-side without any spaces in between
the selections completely break down.
This is not a common occurrance AFAIK so I'll put this on the backburner
for now; there are more important features needed first.